+++ packages/turf-line-to-polygon/test.js
@@ -1,10 +1,10 @@
-import test from 'tape';
 import fs from 'fs';
+import test from 'tape';
 import path from 'path';
 import load from 'load-json-file';
 import write from 'write-json-file';
 import { point, lineString } from '@turf/helpers';
-import lineStringToPolygon from '.';
+import lineToPolygon from '.';
 
 const directories = {
     in: path.join(__dirname, 'test', 'in') + path.sep,
@@ -24,7 +24,7 @@ test('turf-linestring-to-polygon', t => {
     for (const {name, filename, geojson} of fixtures) {
         let {autoComplete, properties, orderCoords} = geojson.properties || {};
         properties = properties || {stroke: '#F0F', 'stroke-width': 6};
-        const results = lineStringToPolygon(geojson, {
+        const results = lineToPolygon(geojson, {
             properties: properties,
             autoComplete: autoComplete,
             orderCoords: orderCoords
@@ -34,8 +34,8 @@ test('turf-linestring-to-polygon', t => {
         t.deepEqual(load.sync(directories.out + filename), results, name);
     }
     // Handle Errors
-    t.throws(() => lineStringToPolygon(point([10, 5])), 'throws - invalid geometry');
-    t.throws(() => lineStringToPolygon(lineString([])), 'throws - empty coordinates');
-    t.assert(lineStringToPolygon(lineString([[10, 5], [20, 10], [30, 20]]), {autocomplete: false}), 'is valid - autoComplete=false');
+    t.throws(() => lineToPolygon(point([10, 5])), 'throws - invalid geometry');
+    t.throws(() => lineToPolygon(lineString([])), 'throws - empty coordinates');
+    t.assert(lineToPolygon(lineString([[10, 5], [20, 10], [30, 20]]), {autocomplete: false}), 'is valid - autoComplete=false');
     t.end();
 });
