+++ packages/turf-line-to-polygon/test/in/collection-linestring.geojson
