+++ packages/turf-line-to-polygon/test/in/geometry-linestring.geojson
