+++ packages/turf-line-to-polygon/test/in/linestring-incomplete.geojson
