+++ packages/turf-line-to-polygon/test/in/linestring.geojson
