+++ packages/turf-line-to-polygon/test/in/linestrings-to-multipolygons.geojson
