+++ packages/turf-line-to-polygon/test/in/multi-linestring-nested2.geojson
