+++ packages/turf-line-to-polygon/test/in/multi-linestring-outer-ring-middle-position.geojson
