+++ packages/turf-line-to-polygon/test/in/multi-linestring-with-hole.geojson
