+++ packages/turf-line-to-polygon/test/in/multi-linestrings-nested.geojson
