+++ packages/turf-line-to-polygon/test/in/multi-linestrings-outer-doughnut.geojson
