+++ packages/turf-line-to-polygon/test/in/multi-linestrings-with-holes.geojson
