+++ packages/turf/index.d.ts
@@ -62,8 +62,8 @@ export {default as greatCircle} from '@turf/great-circle';
 export {default as lineSegment} from '@turf/line-segment';
 export {default as lineSplit} from '@turf/line-split';
 export {default as lineArc} from '@turf/line-arc';
-export {default as polygonToLineString} from '@turf/polygon-to-linestring';
-export {default as lineStringToPolygon} from '@turf/linestring-to-polygon';
+export {default as polygonToLine} from '@turf/polygon-to-line';
+export {default as lineToPolygon} from '@turf/line-to-polygon';
 export {default as bboxClip} from '@turf/bbox-clip';
 export {default as lineOverlap} from '@turf/line-overlap';
 export {default as sector} from '@turf/sector';
