+++ packages/turf/package.json
@@ -150,7 +150,7 @@
     "@turf/line-slice": "*",
     "@turf/line-slice-along": "*",
     "@turf/line-split": "*",
-    "@turf/linestring-to-polygon": "*",
+    "@turf/line-to-polygon": "*",
     "@turf/mask": "*",
     "@turf/meta": "*",
     "@turf/midpoint": "*",
@@ -162,7 +162,7 @@
     "@turf/point-on-surface": "*",
     "@turf/point-to-line-distance": "*",
     "@turf/polygon-tangents": "*",
-    "@turf/polygon-to-linestring": "*",
+    "@turf/polygon-to-line": "*",
     "@turf/polygonize": "*",
     "@turf/projection": "*",
     "@turf/random": "*",
